Port scaffold to oclif
